Cosmetic change - changed position of global variables to make them a bit easier to find.
Added lists of models to be executed for IF and FC types of data.
Right now, these are the same lists, since we model FC data by
converting it to IF data and then applying IF models.

Also added as global items lists that will track models that were
successfully executed and those that weren’t.

Fixed a new defect in the renderDataTable call that output an extra
column at the end of the table displaying raw data or trend test
results.  Don’t know how this fault was introduced, but it happened
after I updated my R packages. on 8/11/2015.