See change details below.
Changes in this commit:

Model_specifications.R

Added two items to a model’s specification - “failcount” and
“numfailsparm”.  The first new spec can have values of “finite” or
“infinite”, and specifies whether a model is a finite failures or
infinite failures model.  For finite-failures models, the specification
“numfailsparm” identifies which model parameter in the specification
item “params” gives the total number of failures expected to be
observed.  For infinite-failures models, this new specification item is
set to 0.

server.R

	1.	Commented out print statement in reactive definition of data_global.
	2.	Reinserted reactive data items DTP_height and DTPranges to make
data and trend plot height track the plot width.
	3.	Renamed “distPlot” to “DataAndTrendPlot”.
	4.	Added global variables at start of file.
	5.	Added global “constants” including lists of applicable models for
IF and FC models, and colors associated with specific models.
	6.	Added reactive item “output$DataSubsetError” which monitors the
length of the modeling data range.  If there are less than 5
observations in the modeling data range, the sliders are forced to
include 5 observations.
	7.	Changed options in renderDataTable displaying raw data and trend
tests to include an option of showing all rows at once.
	8.	Added download handler “output$saveDataOrTrend” to save plots of
raw data and trend tests.

ui.R

	1.	Renamed “distPlot” to “DataAndTrendPlot”.
	2.	Added numeric input “modelRelIntvl” to let users choose the mission
time for which reliability will be estimated.
	3.	Added selectInput “modelsToRun” to let users choose which models
will be run.  This is dynamically populated depending on the type of
the data file being used (IF or FC).

DataFormat.R

	1.	Added functions “FCFrame_to_IFFrame” and “”.

Replaced current Plot_Raw_Data with the version I had developed in my
own branch.