server.R: Two corrections - 1) set default value of mission length to most recent non-zero IF time in FT data.  2) when calling model_result_table, pass the length of the first column of modeled data.
PlotModelResults.R - renamed intermediate buffer which holds dynamic
reliability growth estimates and predictions.  Changed title of
reliability growth plot to add the length of the mission for which
reliability growth is being estimated.

ModelResultTable.R - made reliability growth a dynamic column as it is
for the reliability growth plot.  When a user changes the length of the
mission, that change is reflected in updated values for the reliability
growth columns displayed.  The table is also now displayed as
characters, so that value of NA, Inf, and NaN are shown.

DSS_BM_FT.R - corrected apparent typo in expression for computing MTTFs.