RunModels.R: Using numDeriv library’s “hessian” function to compute hessian from likelihood function and then compute confidence bounds. Replaces use of “optim” function.  NOTE: it is necessary to specify the “numDeriv” library in the call to “hessian”, since the rootSolve library also has a hessian function.
Also modified call to models’ likelihood functions to be compatible
with how they were rewritten to be compatible with the numDeriv
“hessian” function.

GM_BM.R, GO_BM_FT.R, JM_BM.R: Rewrote likelihood functions to be
compatible with use of the “hessian” function in the numDeriv library.

server.R: Modified call to models’ likelihood functions to be
compatible with how they were rewritten to be compatible with the
numDeriv “hessian” function.