RunModels.R: Numerous changes to compute model results for high and low confidence bounds as well as MLEs.  Changes structure of model results table, and changes the structure of list of successful and unsuccessful models.  Each of these lists now contains successful and unsuccessful models for MLEs as well as high and low confidence values.
There is now a new situation for finite-failure models in which the low
confidence bound value for the expected number of failures is less than
the number of failures actually observed.  We need to add code to
handle this new situation.

server.R:  Recognizes structural changes to list of successful and
unsuccessful models.